Move `GenerateTestApplicationsForProvider` out of Vendor API namespace
This class is now also used in the support interface.
